[UPDATE]: Size of wndpool_t is reduced

+++ include/io/message/message.h
@@ -11,7 +11,7 @@ typedef uint32_t (*mc_time_now_us_fn)();
 typedef void     (*mc_msg_on_receive_fn)(const void* const data, uint32_t size);
 
 
-mc_msg_t* mc_msg_new(mc_msg_read_fn read_fn, mc_msg_write_fn write_fn, uint32_t window_size, uint32_t capacity, mc_msg_on_receive_fn on_receive, mc_time_now_us_fn now_us);
+mc_msg_t* mc_msg_new(mc_msg_read_fn read_fn, mc_msg_write_fn write_fn, uint32_t window_size, uint8_t capacity, mc_msg_on_receive_fn on_receive, mc_time_now_us_fn now_us);
 void      mc_msg_free(mc_msg_t** const msg);
 mc_result mc_msg_clear(mc_msg_t* const msg);
 uint32_t  mc_msg_get_capacity(mc_msg_t* const msg);
